+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   Although removed from openeo-processes 2.0.0-rc1, support for `load_result`, `predict_random_forest` and `load_ml_model`
   is preserved but deprecated. ([#424](https://github.com/Open-EO/openeo-python-client/issues/424))
 - Show more informative error message on `403 Forbidden` errors from CDSE firewall ([#512](https://github.com/Open-EO/openeo-python-client/issues/512))
+- Handle API error responses more strict and avoid hiding possibly important information in JSON-formatted but non-compliant error responses.
 
 ### Removed